add: error middleware change: separate todos route fix: todocontroller throw error to error middleware, todoservice no more catch error, todorepository throw error to error middleware

+++ src/service/TodoService.js
@@ -1,69 +1,43 @@
-import { InternalServerError, NotFoundError } from "../exception/error.js"
-import { sanitize } from "../helper/regex.js"
+import { BadRequestError, InternalServerError, NotFoundError } from "../exception/error.js"
+import { isId, sanitize } from "../helper/regex.js"
 
 const TodoService = (TodoRepository) => ({
     FindAll: async () => {
-        try {
-            return await TodoRepository.FindAll()
-        } catch (e) {
-            throw InternalServerError(e.message)
-        }
+        return await TodoRepository.FindAll()
     },
 
     FindOne: async (id) => {
-        try {
-            const find = await TodoRepository.FindOne(id)
-            if (!find) throw NotFoundError()
-            return find
-        } catch (e) {
-            if (e.code !== 404) {
-                throw InternalServerError(e.message)
-            }
-            throw e
-        }
+        if (!isId(id)) throw BadRequestError('ID must be number')
+        const find = await TodoRepository.FindOne(id)
+        if (!find) throw NotFoundError()
+        return find
     },
 
     Create: async (todo) => {
-        try {
-            todo.name = sanitize(todo.name)
-            todo.status = sanitize(todo.status)
-            todo.date = new Date().toISOString()
-            return await TodoRepository.Create(todo)
-        } catch (e) {
-            throw InternalServerError(e.message)
-        }
+        todo.name = sanitize(todo.name)
+        todo.status = sanitize(todo.status)
+        todo.date = new Date().toISOString()
+        return await TodoRepository.Create(todo)
     },
 
     Update: async (id, todo) => {
-        try {
-            todo.name = sanitize(todo.name)
-            todo.status = sanitize(todo.status)
-            const find = await TodoRepository.FindOne(id)
-            if (!find) throw NotFoundError()
-            todo.name ? todo.name : todo.name = find.name
-            todo.status ? todo.status : todo.status = find.status
-            todo.date = find.date
-            const result = await TodoRepository.Update(find.id, todo)
-            return result
-        } catch (e) {
-            if (e.code !== 404) {
-                throw InternalServerError(e.message)
-            }
-            throw e
-        }
+        if (!isId(id)) throw BadRequestError('ID must be number')
+        todo.name = sanitize(todo.name)
+        todo.status = sanitize(todo.status)
+        const find = await TodoRepository.FindOne(id)
+        if (!find) throw NotFoundError()
+        todo.name ? todo.name : todo.name = find.name
+        todo.status ? todo.status : todo.status = find.status
+        todo.date = find.date
+        const result = await TodoRepository.Update(find.id, todo)
+        return result
     },
 
     Delete: async (id) => {
-        try {
-            const find = await TodoRepository.FindOne(id)
-            if (!find) throw NotFoundError()
-            return await TodoRepository.Delete(id)
-        } catch (e) {
-            if (e.code !== 404) {
-                throw InternalServerError(e.message)
-            }
-            throw e
-        }
+        if (!isId(id)) throw BadRequestError('ID must be number')
+        const find = await TodoRepository.FindOne(id)
+        if (!find) throw NotFoundError()
+        return await TodoRepository.Delete(id)
     }
 })
